Added automatic covid results
