+++ start-dev.sh
@@ -37,6 +37,7 @@ services:
       - DISCORD_COMMITTEE_SERVER=${COMMITTEE_SERVER}
       - DISCORD_COMMITTEE_CHANNEL=${COMMITTEE_CHANNEL}
       - DISCORD_PUBLIC_WELCOME=${WELCOME_MESSAGES}
+      - DISCORD_PUBLIC_CORONA=${PUBLIC_GENERAL}
       - SENDGRID_TOKEN=${SENDGRID}
     volumes:
       - ${WD}:/bot
